Add custom, patched endpoint
There's a bug either on the deepstream server or in autobahn's wrapper
factory. When using the wrapper, autobahn will try to negotiate a
websocket subprotocol. Deepstream will reflect that response. Autobahn
won't parse the returned list correctly. (Is the server supposed to
choose only one in the websocket spec?) I spent a lot of time trying to
work around this and patch it.